# Micro optimize LazyInternalScopeFactory